login moved to index page and edited to only have log in button
