+++ common/utils.js
@@ -7,3 +7,11 @@ export const findById = (array, id) => {
         }
     }
 };
+
+export function makeUser(formData) {
+    const user = {
+        username: formData.get('username'),
+        password: formData.get('password')
+    };
+    return user;
+}
