+++ index.html
@@ -6,11 +6,29 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="style.css">
-    <title>Document</title>
+    <title>Portland Happy Hour</title>
 </head>
 
 <body>
-
+<body>
+    <header>
+        <h1>Portland Happy Hour</h1>
+        <h2>Don't just stand there, explore the city!</h2>
+    </header>
+    <main>
+        <form id='login'>
+            <label>
+                <span>Username:</span>
+                <input type='text' id='username' name='username'>
+            </label>
+            <label>
+                <span>Password:</span>
+                <input type='password' id='password' name='password'>
+            </label>
+            <button type='submit' id='login-button'>Login</button>
+        </form>
+    </main>
+    <script src='/data/login.js' type='module'></script>
     <script src="app.js"></script>
 </body>
