--- login/login.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <link href='./login.css' rel='stylesheet'>
-</head>
-<body>
-    <header>
-        <h1>Portland Happy Hour</h1>
-        <h2>Don't just stand there, explore the city!</h2>
-    </header>
-    <main>
-        <form id='login'>
-            <label>
-                <span>Username:</span>
-                <input type='text' id='username' name='username'>
-            </label>
-            <label>
-                <span>Password:</span>
-                <input type='password' id='password' name='password'>
-            </label>
-            <button type='submit' id='register'>Register</button>
-            <button type='submit' id='login-button'>Login</button>
-        </form>
-    </main>
-    <script src='login.js' type='module'></script>
-</body>
-</html>
