--- login/makeuser.js
@@ -1,7 +1,0 @@
-export function makeUser(formData) {
-    const user = {
-        username: formData.get('username'),
-        password: formData.get('password')
-    };
-    return user;
-}
