+++ map/map.html
@@ -10,7 +10,7 @@
 <body>
     <figure>
     </figure>
-    THIS IS THE MAP HTML PAGE
+    
 </body>
 <script src="map.js"></script>
 </html>
