+++ style.css
@@ -1,0 +1,33 @@
+body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    background-color: salmon;
+    width: 700px;
+    height: 100%;
+    margin: 0 auto;
+}
+html {
+    background-color: white;
+    height: 100%;
+}
+
+form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
+input {
+    display: flex;
+}
+button {
+    margin-top: 10px;
+}
+header {
+    margin-top: 3%;
+}
+input {
+    border: 0px 2px 0px 0px;
+}
